Restructure api files
