+++ 03_Modules/Configuration/src/index.ts
@@ -3,8 +3,8 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { ConfigurationModuleApi as ConfigurationDataApi } from './module/DataApi';
-export { ConfigurationModuleApi as ConfigurationOcpp201Api } from './module/Ocpp201Api';
-export { ConfigurationModuleApi as ConfigurationOcpp16Api } from './module/Ocpp16Api';
+export { ConfigurationDataApi } from './module/DataApi';
+export { ConfigurationOcpp201Api } from './module/Ocpp201Api';
+export { ConfigurationOcpp16Api } from './module/Ocpp16Api';
 export { IConfigurationModuleApi } from './module/interface';
 export { ConfigurationModule } from './module/module';
