+++ 03_Modules/Monitoring/src/module/Ocpp201Api.ts
@@ -7,36 +7,21 @@ import { ILogObj, Logger } from 'tslog';
 import { IMonitoringModuleApi } from './interface';
 import { MonitoringModule } from './module';
 import {
-  Component,
-  CreateOrUpdateVariableAttributeQuerySchema,
-  CreateOrUpdateVariableAttributeQuerystring,
-  sequelize,
-  Variable,
-  VariableAttributeQuerySchema,
-  VariableAttributeQuerystring,
-} from '@citrineos/data';
-import {
   AbstractModuleApi,
-  AsDataEndpoint,
   AsMessageEndpoint,
   CallAction,
-  HttpMethod,
   IMessageConfirmation,
-  OCPP2_0_1_Namespace,
   OCPP2_0_1,
   OCPP2_0_1_CallAction,
   OCPPVersion,
-  ReportDataTypeSchema,
-  OCPP1_6_Namespace,
-  Namespace,
 } from '@citrineos/base';
-import { FastifyInstance, FastifyRequest } from 'fastify';
+import { FastifyInstance } from 'fastify';
 import { getBatches, getSizeOfRequest } from '@citrineos/util';
 
 /**
  * Server API for the Monitoring module.
  */
-export class MonitoringModuleApi
+export class MonitoringOcpp201Api
   extends AbstractModuleApi<MonitoringModule>
   implements IMonitoringModuleApi
 {
@@ -55,12 +40,8 @@ export class MonitoringModuleApi
     server: FastifyInstance,
     logger?: Logger<ILogObj>,
   ) {
-    super(monitoringModule, server, logger);
+    super(monitoringModule, server, OCPPVersion.OCPP2_0_1, logger);
   }
-
-  /**
-   * Message Endpoints
-   */
 
   @AsMessageEndpoint(
     OCPP2_0_1_CallAction.SetVariableMonitoring,
@@ -387,89 +368,6 @@ export class MonitoringModuleApi
   }
 
   /**
-   * Data Endpoints
-   */
-
-  @AsDataEndpoint(
-    OCPP2_0_1_Namespace.VariableAttributeType,
-    HttpMethod.Put,
-    CreateOrUpdateVariableAttributeQuerySchema,
-    ReportDataTypeSchema,
-  )
-  async putDeviceModelVariables(
-    request: FastifyRequest<{
-      Body: OCPP2_0_1.ReportDataType;
-      Querystring: CreateOrUpdateVariableAttributeQuerystring;
-    }>,
-  ): Promise<sequelize.VariableAttribute[]> {
-    // fill in default values where omitted
-    for (const variableAttr of request.body.variableAttribute) {
-      if (!variableAttr.mutability) {
-        variableAttr.mutability = OCPP2_0_1.MutabilityEnumType.ReadWrite;
-      }
-    }
-    const timestamp = new Date().toISOString();
-    const variableAttributes =
-      await this._module.deviceModelRepository.createOrUpdateDeviceModelByStationId(
-        request.body,
-        request.query.stationId,
-        timestamp,
-      );
-
-    if (request.query.setOnCharger) {
-      // Mark them as Accepted (set on the charger outside of OCPP communication)
-      for (let variableAttribute of variableAttributes) {
-        variableAttribute = await variableAttribute.reload({
-          include: [Variable, Component],
-        });
-        await this._module.deviceModelRepository.updateResultByStationId(
-          {
-            attributeType: variableAttribute.type,
-            attributeStatus: OCPP2_0_1.SetVariableStatusEnumType.Accepted,
-            attributeStatusInfo: { reasonCode: 'SetOnCharger' },
-            component: variableAttribute.component,
-            variable: variableAttribute.variable,
-          },
-          request.query.stationId,
-          timestamp,
-        );
-      }
-    }
-    return variableAttributes;
-  }
-
-  @AsDataEndpoint(
-    OCPP2_0_1_Namespace.VariableAttributeType,
-    HttpMethod.Get,
-    VariableAttributeQuerySchema,
-  )
-  getDeviceModelVariables(
-    request: FastifyRequest<{ Querystring: VariableAttributeQuerystring }>,
-  ): Promise<sequelize.VariableAttribute[] | undefined> {
-    return this._module.deviceModelRepository.readAllByQuerystring(
-      request.query,
-    );
-  }
-
-  @AsDataEndpoint(
-    OCPP2_0_1_Namespace.VariableAttributeType,
-    HttpMethod.Delete,
-    VariableAttributeQuerySchema,
-  )
-  deleteDeviceModelVariables(
-    request: FastifyRequest<{ Querystring: VariableAttributeQuerystring }>,
-  ): Promise<string> {
-    return this._module.deviceModelRepository
-      .deleteAllByQuerystring(request.query)
-      .then(
-        (deletedCount) =>
-          deletedCount.toString() +
-          ' rows successfully deleted from ' +
-          OCPP2_0_1_Namespace.VariableAttributeType,
-      );
-  }
-
-  /**
    * Processes data in batches and sends them to the specified OCPP action.
    *
    * @param {string} stationId - The station's identifier.
@@ -536,20 +434,5 @@ export class MonitoringModuleApi
     const endpointPrefix =
       this._module.config.modules.monitoring.endpointPrefix;
     return super._toMessagePath(input, endpointPrefix);
-  }
-
-  /**
-   * Overrides superclass method to generate the URL path based on the input {@link Namespace}
-   * and the module's endpoint prefix configuration.
-   *
-   * @param {Namespace} input - The input {@link Namespace}.
-   * @return {string} - The generated URL path.
-   */
-  protected _toDataPath(
-    input: OCPP2_0_1_Namespace | OCPP1_6_Namespace | Namespace,
-  ): string {
-    const endpointPrefix =
-      this._module.config.modules.monitoring.endpointPrefix;
-    return super._toDataPath(input, endpointPrefix);
   }
 }
