+++ 03_Modules/OcppRouter/src/index.ts
@@ -3,7 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { AdminApi } from './module/api';
+export { AdminApi } from './module/DataApi';
 export { IAdminApi } from './module/interface';
 export { MessageRouterImpl } from './module/router';
 export { WebhookDispatcher } from './module/webhook.dispatcher';
