+++ 03_Modules/Tenant/src/module/DataApi.ts
@@ -18,7 +18,7 @@ import { ILogObj, Logger } from 'tslog';
 /**
  * Server API for the Tenant module.
  */
-export class TenantModuleApi
+export class TenantDataApi
   extends AbstractModuleApi<TenantModule>
   implements ITenantModuleApi
 {
@@ -35,25 +35,14 @@ export class TenantModuleApi
     server: FastifyInstance,
     logger?: Logger<ILogObj>,
   ) {
-    super(tenantModule, server, logger);
+    super(tenantModule, server, null, logger);
   }
 
   /**
-   * Overrides superclass method to generate the URL path based on the input {@link CallAction} and the module's endpoint prefix configuration.
+   * Overrides superclass method to generate the URL path based on the input {@link Namespace}
+   * and the module's endpoint prefix configuration.
    *
-   * @param {CallAction} input - The input {@link CallAction}.
-   * @return {string} - The generated URL path.
-   */
-  protected _toMessagePath(input: CallAction): string {
-    const endpointPrefix = this._module.config.modules.tenant.endpointPrefix;
-    return super._toMessagePath(input, endpointPrefix);
-  }
-
-  /**
-   * Overrides superclass method to generate the URL path based on the input ({@link OCPP2_0_1_Namespace},
-   * {@link OCPP1_6_Namespace} or {@link Namespace}) and the module's endpoint prefix configuration.
-   *
-   * @param {CallAction} input - The input {@link OCPP2_0_1_Namespace}, {@link OCPP1_6_Namespace} or {@link Namespace}.
+   * @param {Namespace} input - The input {@link Namespace}.
    * @return {string} - The generated URL path.
    */
   protected _toDataPath(
