+++ 03_Modules/Transactions/src/index.ts
@@ -3,6 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-export { TransactionsModuleApi } from './module/api';
+export { TransactionsOcpp201Api } from './module/Ocpp201Api';
+export { TransactionsDataApi } from './module/DataApi';
 export { ITransactionsModuleApi } from './module/interface';
 export { TransactionsModule } from './module/module';
