+++ 03_Modules/Transactions/src/module/DataApi.ts
@@ -17,14 +17,9 @@ import { TransactionsModule } from './module';
 import {
   AbstractModuleApi,
   AsDataEndpoint,
-  AsMessageEndpoint,
-  CallAction,
   HttpMethod,
-  IMessageConfirmation,
   OCPP2_0_1_Namespace,
   OCPP2_0_1,
-  OCPP2_0_1_CallAction,
-  OCPPVersion,
   OCPP1_6_Namespace,
   Namespace,
 } from '@citrineos/base';
@@ -35,7 +30,7 @@ import { plainToInstance } from 'class-transformer';
 /**
  * Server API for the transaction module.
  */
-export class TransactionsModuleApi
+export class TransactionsDataApi
   extends AbstractModuleApi<TransactionsModule>
   implements ITransactionsModuleApi
 {
@@ -51,61 +46,8 @@ export class TransactionsModuleApi
     server: FastifyInstance,
     logger?: Logger<ILogObj>,
   ) {
-    super(transactionModule, server, logger);
+    super(transactionModule, server, null, logger);
   }
-
-  /**
-   * Message Endpoint Methods
-   */
-  @AsMessageEndpoint(
-    OCPP2_0_1_CallAction.CostUpdated,
-    OCPP2_0_1.CostUpdatedRequestSchema,
-  )
-  async costUpdated(
-    identifier: string[],
-    tenantId: string,
-    request: OCPP2_0_1.CostUpdatedRequest,
-    callbackUrl?: string,
-  ): Promise<IMessageConfirmation[]> {
-    const results = identifier.map((id) =>
-      this._module.sendCall(
-        id,
-        tenantId,
-        OCPPVersion.OCPP2_0_1,
-        OCPP2_0_1_CallAction.CostUpdated,
-        request,
-        callbackUrl,
-      ),
-    );
-    return Promise.all(results);
-  }
-
-  @AsMessageEndpoint(
-    OCPP2_0_1_CallAction.GetTransactionStatus,
-    OCPP2_0_1.GetTransactionStatusRequestSchema,
-  )
-  getTransactionStatus(
-    identifier: string[],
-    tenantId: string,
-    request: OCPP2_0_1.GetTransactionStatusRequest,
-    callbackUrl?: string,
-  ): Promise<IMessageConfirmation[]> {
-    const results = identifier.map((id) =>
-      this._module.sendCall(
-        id,
-        tenantId,
-        OCPPVersion.OCPP2_0_1,
-        OCPP2_0_1_CallAction.GetTransactionStatus,
-        request,
-        callbackUrl,
-      ),
-    );
-    return Promise.all(results);
-  }
-
-  /**
-   * Data Endpoint Methods
-   */
 
   @AsDataEndpoint(
     OCPP2_0_1_Namespace.TransactionType,
@@ -120,9 +62,6 @@ export class TransactionsModuleApi
       request.query.transactionId,
     );
   }
-
-  // TODO: Determine how to implement readAllTransactionsByStationIdAndChargingStates as a GET...
-  // TODO: Determine how to implement existsActiveTransactionByIdToken as a GET...
 
   @AsDataEndpoint(
     OCPP2_0_1_Namespace.Tariff,
@@ -164,19 +103,6 @@ export class TransactionsModuleApi
           ' rows successfully deleted from ' +
           OCPP2_0_1_Namespace.Tariff,
       );
-  }
-
-  /**
-   * Overrides superclass method to generate the URL path based on the input {@link CallAction}
-   * and the module's endpoint prefix configuration.
-   *
-   * @param {CallAction} input - The input {@link CallAction}.
-   * @return {string} - The generated URL path.
-   */
-  protected _toMessagePath(input: CallAction): string {
-    const endpointPrefix =
-      this._module.config.modules.transactions.endpointPrefix;
-    return super._toMessagePath(input, endpointPrefix);
   }
 
   /**
