+++ 03_Modules/Transactions/src/module/interface.ts
@@ -3,19 +3,7 @@
 //
 // SPDX-License-Identifier: Apache 2.0
 
-import {
-  OCPP2_0_1,
-  IMessageConfirmation,
-} from '@citrineos/base';
-
 /**
  * Interface for the transaction module.
  */
-export interface ITransactionsModuleApi {
-  getTransactionStatus(
-    identifier: string[],
-    tenantId: string,
-    request: OCPP2_0_1.GetTransactionStatusRequest,
-    callbackUrl?: string,
-  ): Promise<IMessageConfirmation[]>;
-}
+export interface ITransactionsModuleApi {}
